batch: ShouldSeal compares in ulong space (no int-cast wrap)
`var blocksInBatch = builder.Batch.LastBlock - builder.Batch.FirstBlock
+ 1;` is ulong. The previous `(int)blocksInBatch >=
_settings.MaxBlocksPerBatch` cast wrapped when blocksInBatch >
int.MaxValue (~2.1B), making the seal-by-blocks check silently never
fire — an ultra-long-lived sequencer would accumulate unboundedly.

Now compares in ulong space:
  blocksInBatch >= (ulong)_settings.MaxBlocksPerBatch

iter-191's ctor validation ensures MaxBlocksPerBatch > 0 so the cast
direction is safe.

No pinning test — constructing a 2.1B-block batch in a unit test isn't
practical; the fix is a 1-character symmetric-cast change.

Cumulative: 491 tests / 27 projects.

+++ src/Neo.Plugins.L2Batch/BatchSealer.cs
@@ -109,8 +109,13 @@ public sealed class BatchSealer
 
     private bool ShouldSeal(BatchBuilder builder)
     {
+        // Compare in ulong space to avoid the int-cast wrap. The previous (int)cast
+        // wrapped when blocksInBatch > int.MaxValue (~2.1B), making the seal-by-blocks
+        // check silently never fire — ultra-long-lived sequencers would accumulate
+        // unboundedly. iter-191 BatchSealer ctor validates MaxBlocksPerBatch > 0, so
+        // the unchecked-cast is safe in the other direction.
         var blocksInBatch = builder.Batch.LastBlock - builder.Batch.FirstBlock + 1;
-        if ((int)blocksInBatch >= _settings.MaxBlocksPerBatch) return true;
+        if (blocksInBatch >= (ulong)_settings.MaxBlocksPerBatch) return true;
         if (builder.Batch.TransactionCount >= _settings.MaxTransactionsPerBatch) return true;
         var ageMs = _nowUtcMillis() - _batchStartedAtUtcMillis;
         if (ageMs >= _settings.MaxBatchAgeMillis) return true;
